move not found page into pages folder
