+++ frontend/src/Pages/NotFoundPage.tsx
@@ -1,4 +1,4 @@
 import react from 'react';
-import { Page } from './Page';
+import { Page } from '../Page';
 
 export const NotFoundPage = () => <Page title="Page Not Found">{null}</Page>;
